+++ test/spec/spec_test.exs
@@ -5,7 +5,7 @@ defmodule Oasis.Spec.SpecTest do
 
   test "parse basic info from yaml and json" do
     file_path = Path.join([@dir, "basic.yaml"])
-    %ExJsonSchema.Schema.Root{schema: schema} = Oasis.Spec.read(file_path)
+    %Oasis.Spec.Document{schema: schema} = Oasis.Spec.read(file_path)
 
     assert schema["openapi"] == "3.1.0"
 
@@ -16,7 +16,7 @@ defmodule Oasis.Spec.SpecTest do
     assert length(schema["servers"]) == 3
 
     file_path = Path.join([@dir, "basic.json"])
-    %ExJsonSchema.Schema.Root{schema: schema_from_json} = Oasis.Spec.read(file_path)
+    %Oasis.Spec.Document{schema: schema_from_json} = Oasis.Spec.read(file_path)
 
     assert schema_from_json == schema
   end
@@ -50,16 +50,16 @@ defmodule Oasis.Spec.SpecTest do
 
   test "parse post requestBodies" do
     file_path = Path.join([@dir, "basic.yaml"])
-    %ExJsonSchema.Schema.Root{schema: schema} = Oasis.Spec.read(file_path)
+    %Oasis.Spec.Document{schema: schema} = Oasis.Spec.read(file_path)
 
     request_body_of_refresh_token = schema["paths"]["/refresh_token"]["post"]["requestBody"]
     assert request_body_of_refresh_token["required"] == true
 
     schema = request_body_of_refresh_token["content"]["application/json"]["schema"]
-    schema = ExJsonSchema.Schema.resolve(schema)
+    schema = Oasis.JSONSchema.compile!(schema)
 
-    assert ExJsonSchema.Validator.validate(schema, %{"refresh_token" => "123"}) == :ok
+    assert Oasis.JSONSchema.validate(schema, %{"refresh_token" => "123"}) == :ok
 
-    assert {:error, _} = ExJsonSchema.Validator.validate(schema, %{})
+    assert {:error, _} = Oasis.JSONSchema.validate(schema, %{})
   end
 end